Move LessonPrompts into pages/lessons/components/
