+++ src/pages/lessons/MainLessonView.tsx
@@ -11,9 +11,9 @@ import TypedText from "../../components/TypedText";
 import Scores from "../../components/Scores";
 import StrokeTip from "./components/StrokeTip";
 import UserSettings from "./components/UserSettings/UserSettings";
-import AussieDictPrompt from "../../components/LessonPrompts/AussieDictPrompt";
-import SedSaidPrompt from "../../components/LessonPrompts/SedSaidPrompt";
-import WordBoundaryErrorPrompt from "../../components/LessonPrompts/WordBoundaryErrorPrompt";
+import AussieDictPrompt from "./components/LessonPrompts/AussieDictPrompt";
+import SedSaidPrompt from "./components/LessonPrompts/SedSaidPrompt";
+import WordBoundaryErrorPrompt from "./components/LessonPrompts/WordBoundaryErrorPrompt";
 import PunctuationDescription from "./components/PunctuationDescription";
 import LessonFinePrintFooter from "./components/LessonFinePrintFooter";
 import { useAnnouncerApi } from "../../components/Announcer/useAnnouncer";
